pyproject.toml
[project]
name = "nero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.22.0",
    "chromadb>=1.3.5",
    "einops>=0.8.1",
    "huggingface-hub>=0.36.0",
    "mlx>=0.29.3",
    "mlx-lm>=0.28.4",
    "mlx-vlm>=0.0.11",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
